+++ app/src/main/java/woowacourse/movie/adapter/viewholder/AdViewHolder.kt
@@ -3,7 +3,7 @@ package woowacourse.movie.adapter.viewholder
 import android.view.View
 import android.widget.ImageView
 import woowacourse.movie.R
-import woowacourse.movie.adapter.ItemClickListener
+import woowacourse.movie.listener.ItemClickListener
 import woowacourse.movie.model.MovieListItem
 
 class AdViewHolder(view: View) :
