+++ app/src/main/java/woowacourse/movie/adapter/viewholder/CustomViewHolder.kt
@@ -2,7 +2,7 @@ package woowacourse.movie.adapter.viewholder
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
-import woowacourse.movie.adapter.ItemClickListener
+import woowacourse.movie.listener.ItemClickListener
 import woowacourse.movie.model.MovieListItem
 
 abstract class CustomViewHolder(view: View) : RecyclerView.ViewHolder(view) {
